% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DD_model_average.R
\name{add_average_models}
\alias{add_average_models}
\title{Add average models to a list of fitted models}
\usage{
add_average_models(fitted_models, variance_type = "constant")
}
\arguments{
\item{fitted_models}{The list of fitted models.}

\item{variance_type}{variance type, possible values are: constant, nonconstant, infer.}
}
\value{
The updated list of fitted models with the average model added.
}
\description{
This function adds average models to a list of fitted models based on the model averaging approach described
in "A brief guide to model selection, multimodel inference and model averaging in behavioural ecology using Akaike’s
information criterion" by Matthew R. E. Symonds and Adnan Moussalli. The average model is computed using the provided
models and added to the list under the name "average". The average model is fitted using the data from the first model
in the list. The function returns the updated list of fitted models.
}
