% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PLOT_statistics.R
\name{aggregate_rows_time}
\alias{aggregate_rows_time}
\title{Aggregate Rows by Time and Other Features}
\usage{
aggregate_rows_time(
  mod_stats,
  gen_feat = "Feature",
  first_feat = "SACRI_PERIOD",
  group_by = c("Experiment", "DILI"),
  filter_column = NULL,
  filter_by = list(c("acetaminophen", "bucetin"))
)
}
\arguments{
\item{mod_stats}{A data frame containing model statistics.}

\item{gen_feat}{The general feature to aggregate.}

\item{first_feat}{The first feature used in aggregation.}

\item{group_by}{A character vector specifying the grouping features.}

\item{filter_column}{A character vector specifying the columns to filter.}

\item{filter_by}{A list of character vectors containing filter values for each column.}
}
\value{
Returns a ggplot2 plot of aggregated data or an empty plot if conditions are not met.
}
\description{
Aggregates rows of a data frame based on time and other specified features.
}
