% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DD_point_of_departure.R
\name{bmr_factor}
\alias{bmr_factor}
\title{Calculate the benchmark response level}
\usage{
bmr_factor(risk_factor = 0.1, increase = TRUE, background_level = 0.01)
}
\arguments{
\item{risk_factor}{The risk factor used to determine the benchmark response level.
Default is 0.1.}

\item{increase}{Logical value indicating whether the response should be increased.
If TRUE, the response is increased; if FALSE, it is decreased. Default is TRUE.}

\item{background_level}{The background level of the response. Default is 0.01.}
}
\value{
The benchmark response level calculated based on the input parameters.
}
\description{
This function calculates the benchmark response level based on the given risk factor,
whether the response should be increased or decreased, and the background level.
The calculation is based on Equation 14 from the paper referenced in the code.
}
