% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/INPUT_FILTERING_fold_change_limma.R
\name{build.model.matrix}
\alias{build.model.matrix}
\title{Build Model Matrix}
\usage{
build.model.matrix(
  pd,
  intercept = -1,
  var.int,
  covariates = NULL,
  verbose = TRUE
)
}
\arguments{
\item{pd}{The phenotype data as a data frame.}

\item{intercept}{The value to be used for the intercept (default: -1).}

\item{var.int}{The variable of interest used in the design matrix.}

\item{covariates}{A character vector specifying the optional covariates to include in the design matrix (default: NULL).}

\item{verbose}{Logical; if TRUE, print the formula used for model.matrix (default: TRUE).}
}
\value{
The design matrix for the linear model.
}
\description{
This function constructs the design matrix for a linear model based on the provided
phenotype data, variable of interest, and optional covariates.
}
