% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MODEL_construct_models.R
\name{build_models}
\alias{build_models}
\title{Build multiple models based on the given model names.}
\usage{
build_models(
  model_names = c("linear", "hill", "power", "poly2", "poly3", "poly4", "poly5", "exp2",
    "exp3", "exp4", "exp5", "llog2", "llog3", "llog4", "llog5", "mm2", "weibul12",
    "weibul13", "weibul14", "weibul22", "weibul23", "weibul24"),
  max_iter = 1024,
  data_type = c("continuous", "binomial"),
  x,
  y
)
}
\arguments{
\item{model_names}{A character vector specifying the model names to build.
Default is all available model names.}

\item{max_iter}{Maximum number of iterations for iterative model fitting.
Default is 1024.}

\item{data_type}{Type of data, either "continuous" or "binomial".
Default is "continuous".}

\item{x}{The predictor variable (independent variable).}

\item{y}{The response variable (dependent variable).}
}
\value{
model_list_result A list containing the specified model objects.
}
\description{
This function constructs a list of model objects based on the specified model
names. The available model families include "linear", "hill", "power", "poly2",
"poly3", "poly4", "poly5", "exp2", "exp3", "exp4", "exp5", "llog2", "llog3",
"llog4", "llog5", "mm2", "weibul12", "weibul13", "weibul14", "weibul22",
"weibul23", and "weibul24".
}
