% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DD_gene_pairs.R
\name{check_edges_in_graph}
\alias{check_edges_in_graph}
\title{Check Presence of Edges in a Graph}
\usage{
check_edges_in_graph(
  graph,
  edge_df,
  feature1 = "Feature 1",
  feature2 = "Feature 2"
)
}
\arguments{
\item{graph}{An \code{igraph} object representing the graph to check against.}

\item{edge_df}{A data frame where the first two columns are node names
for the edges to be checked.}
}
\value{
A data frame with an additional logical column indicating
whether each edge is present in the graph.
}
\description{
This function checks whether edges in a data frame are present in
the specified \code{igraph} object. It uses the \code{filter_existing_nodes()}
helper to ensure only valid nodes are checked, and then marks
presence via a logical column.
}
