% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DD_point_of_departure.R
\name{compute_deviations}
\alias{compute_deviations}
\title{Compute deviations for BMD modeling}
\usage{
compute_deviations(deviation_type = "standard", model, rl = 1.349)
}
\arguments{
\item{deviation_type}{Character string specifying the type of deviation to compute.
Possible values are "standard", "relative", and "absolute".}

\item{model}{The model object used for computation.}

\item{rl}{The relative level used in the deviation calculation. Default is 1.349.}
}
\value{
The computed deviation value based on the specified deviation type.
}
\description{
This function computes deviations for BMD modeling based on the specified deviation type.
Three different types of deviation are available: standard deviation, relative deviation,
and absolute deviation. Refer to the EPA documentation for detailed information on the
deviation types (slide 7) from the provided link.
}
