% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PLOT_compare_experiments.R
\name{compute_gene_frequencies}
\alias{compute_gene_frequencies}
\title{Compute Gene Frequencies and Create Lollipop Plots}
\usage{
compute_gene_frequencies(
  mod_stats,
  th = 0.7,
  top_genes = 100,
  rel_variable = "Experiment",
  group_by = "None",
  split_by = "None"
)
}
\arguments{
\item{mod_stats}{A data frame containing the model statistics and gene information.}

\item{th}{Threshold value for gene percentage. Only genes with a percentage above this threshold will be plotted.}

\item{top_genes}{Maximum number of genes to plot. Default = 100}

\item{rel_variable}{The name of the variable representing the experiments or conditions.}

\item{group_by}{The name of the variable to group the data for generating lollipop plots.}

\item{split_by}{The name of the variable to split the data and generate separate lollipop plots.}
}
\value{
A list containing gene lists, lollipop plots, and matrices for each split-by value or for all data.
}
\description{
This function computes gene frequencies based on the provided model statistics and generates lollipop plots for the top genes based on their frequencies.
}
