% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DD_model_statistics.R
\name{compute_model_statistics}
\alias{compute_model_statistics}
\title{Compute model statistics for fitted models}
\usage{
compute_model_statistics(
  fitted_models,
  other_variables_id_col,
  is_parallel = TRUE,
  nCores = 2
)
}
\arguments{
\item{fitted_models}{A list of fitted models.}

\item{other_variables_id_col}{The name of the column representing the other variables in the model statistics data frame.}

\item{is_parallel}{Whether to compute the statistics in parallel. Default is TRUE.}

\item{nCores}{The number of cores to use for parallel computation. Default is 2.}
}
\value{
A data frame containing the computed model statistics.
}
\description{
This function computes model statistics for a list of fitted models and returns the results as a data frame.
}
