% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/INPUT_create_data_structure_list.R
\name{create_data_structure}
\alias{create_data_structure}
\title{this function convert the data in input into the format required for dose-dependent modelling}
\usage{
create_data_structure(
  experimental_data,
  metadata,
  sample_id_col = "BARCODE",
  dose_id_col = "DOSE",
  other_variables_id_col = c("SACRI_PERIOD"),
  x = "dose",
  y = "expr"
)
}
\arguments{
\item{experimental_data}{a list of dataframe containing experimental data. Each row is a feature (e.g. gene) and each column is a sample}

\item{metadata}{a list of dataframe containing the metadata for the experimental data. Each row is a sample and the columns represent the different variables. A column for dose/concentration is required}

\item{sample_id_col}{a character specifying the name of the column containing the samples id}

\item{dose_id_col}{a character specifying the name of the column containing the doses/concentration}

\item{other_variables_id_col}{a vector of characters specifying the name of the column used to group the data}

\item{x}{a characters specifying the name of the x variable in the model. Default is dose.}

\item{y}{a characters specifying the name of the y variable in the model. Default is expr.}
}
\value{
a dictionary containing the data frame for modelling. Dictionary Keys are n-uple specifying the experiment name, other variables of interests and feature names (e.g. drug, time, gene)
}
\description{
this function convert the data in input into the format required for dose-dependent modelling
}
