% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DD_dose_response_analysis.R
\name{dose_response_analysis}
\alias{dose_response_analysis}
\title{Perform dose-response analysis on a list of models}
\usage{
dose_response_analysis(
  data,
  model_list,
  deviation_type = "standard",
  rl = 1.349,
  variance_type = "constant",
  confidence_interval = 0.95,
  significance_level = 0.05
)
}
\arguments{
\item{data}{The data used for fitting the models.}

\item{model_list}{A list of models to fit to the data.}

\item{deviation_type}{Character string specifying the type of deviation from the fitted
model to use for BMD calculation. Default is "standard". Allowed values are standard and relative}

\item{rl}{The relative level used to calculate the BMD. Default is 1.349.}

\item{variance_type}{Character string specifying the type of variance to use in model
fitting. Default is "constant". Other possible values are "non constant", "model" and "inferred"}

\item{confidence_interval}{The confidence level for the confidence interval. Default is 0.95.}

\item{significance_level}{The significance level for model fitting. Default is 0.05.}
}
\value{
A list of models with additional attributes for BMDL, BMDU, and AC50 values.
Models that fail to fit or encounter an error during estimation are excluded
from the returned list.
}
\description{
This function fits a list of models to the same end-point and performs dose-response
analysis, including estimation of BMD, BMDL, BMDU, and AC50 values.
}
