% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PLOT_compare_experiments.R
\name{ecdf_plots}
\alias{ecdf_plots}
\title{Make ECDF or Histogram Plots for BMD Data}
\usage{
ecdf_plots(
  mod_stats,
  rel_variable = "Experiment",
  group_by = NULL,
  is_group_by_numeric = TRUE,
  other_variables = NULL,
  number_of_column = 2,
  scaling = TRUE,
  filter_column = c("Model"),
  filter_by = list(c("linear")),
  plot_type = "ecdf",
  linewidth = 1.3,
  text_size = 14
)
}
\arguments{
\item{mod_stats}{The model statistics data frame.}

\item{rel_variable}{The column name in \code{mod_stats} representing the experimental condition or treatment (default: "Experiment").}

\item{group_by}{The column name in \code{mod_stats} used to group the data for faceting (default: NULL).}

\item{is_group_by_numeric}{Logical; if TRUE, treat the \code{group_by} variable as numeric (default: TRUE).}

\item{other_variables}{A character vector specifying additional variables for faceting (default: NULL).}

\item{number_of_column}{The number of columns in the facet grid (default: 2).}

\item{scaling}{Logical; if TRUE, scale the BMD values (default: TRUE).}

\item{filter_column}{The column name(s) in \code{mod_stats} to use for filtering (default: "Model").}

\item{filter_by}{A list of character vectors containing filtering criteria for each \code{filter_column} (default: list(c("linear"))).}

\item{plot_type}{The type of plot to create; either "ecdf" for ECDF or "histogram" for histogram (default: "ecdf").}

\item{linewidth}{Integer specifing the width of the lines. Default 1.3}

\item{text_size}{Integer specifing the font size. Default = 14}
}
\value{
A plotly object representing the ECDF or histogram plot.
}
\description{
This function creates ECDF (Empirical Cumulative Distribution Function) or Histogram plots
for the BMD (Benchmark Dose) data in the provided model statistics data frame.
}
