% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ens_human_gene_graph}
\alias{ens_human_gene_graph}
\title{ens_human_gene_graph}
\format{
An object of class \code{igraph} of length 20817.
}
\source{
\url{https://doi.org/10.1093/bib/bbaa417}
}
\usage{
ens_human_gene_graph
}
\description{
Graph whose nodes are Human genes (Ensembl ID). Edges represent protein-protein interactions (PPIs).
}
\keyword{datasets}
