% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ens_mouse_gene_graph_mirna}
\alias{ens_mouse_gene_graph_mirna}
\title{ens_mouse_gene_graph_mirna}
\format{
An object of class \code{igraph} of length 4294.
}
\source{
\url{https://doi.org/10.1093/bib/bbaa417}
}
\usage{
ens_mouse_gene_graph_mirna
}
\description{
Graph whose nodes are Mouse genes (Ensembl ID). Edges represent regulatory interactions between miRNA and genes.
}
\keyword{datasets}
