% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ens_rat_gene_graph_tf}
\alias{ens_rat_gene_graph_tf}
\title{ens_rat_gene_graph_tf}
\format{
An object of class \code{igraph} of length 172.
}
\source{
\url{https://doi.org/10.1093/bib/bbaa417}
}
\usage{
ens_rat_gene_graph_tf
}
\description{
Graph whose nodes are Rat genes (Ensembl ID). Edges represent regulatory interactions between transcription factor (TF) and genes.
}
\keyword{datasets}
