% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ent_human_gene_graph_mirna}
\alias{ent_human_gene_graph_mirna}
\title{ent_human_gene_graph_mirna}
\format{
An object of class \code{igraph} of length 120356739.
}
\source{
\url{https://doi.org/10.1093/bib/bbaa417}
}
\usage{
ent_human_gene_graph_mirna
}
\description{
Graph whose nodes are Human genes (Entrez ID). Edges represent regulatory interactions between miRNA and genes.
}
\keyword{datasets}
