% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ent_rat_gene_graph}
\alias{ent_rat_gene_graph}
\title{ent_rat_gene_graph}
\format{
An object of class \code{igraph} of length 120102993.
}
\source{
\url{https://doi.org/10.1093/bib/bbaa417}
}
\usage{
ent_rat_gene_graph
}
\description{
Graph whose nodes are Rat genes (Entrez ID). Edges represent protein-protein interactions (PPIs).
}
\keyword{datasets}
