% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DD_tpod.R
\name{filter_antimode}
\alias{filter_antimode}
\title{Filter Data Based on Antimode of BMD Distribution}
\usage{
filter_antimode(optimal_models_stats, timepoints_col_variable)
}
\arguments{
\item{optimal_models_stats}{A dataframe containing the BMD values and corresponding timepoints.}

\item{timepoints_col_variable}{A character vector specifying the timepoints column variable to analyze.}
}
\value{
A \strong{list} with two elements:
\itemize{
\item \strong{$before}: Dataframe containing BMD values \strong{before} the antimode.
\item \strong{$after}: Dataframe containing BMD values \strong{after} the antimode but below the 95th percentile.
}
}
\description{
This function filters \code{optimal_models_stats} based on the \strong{antimode} of the BMD distribution.
It splits the data into two groups:
\itemize{
\item \strong{Before Antimode}: BMD values \strong{less than or equal to} the antimode.
\item \strong{After Antimode}: BMD values \strong{greater than} the antimode but \strong{below the 95th percentile}.
}
}
