% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PLOT_statistics.R
\name{filter_df}
\alias{filter_df}
\title{this function filters the dataframe of the models statistic}
\usage{
filter_df(mod_stats, filter_column, filter_by)
}
\arguments{
\item{mod_stats}{the dataframe containing all the statistics}

\item{filter_column}{a vector of dataframe columns to be used for filtering}

\item{filter_by}{a list containing the selected values to be mantained after filtering.
Each position of the list correspond to one of the selected column and contains a vector of admissible values
#The length of filter_column and filter_by must be the same}
}
\value{
the filtered dataframe
}
\description{
this function filters the dataframe of the models statistic
}
