% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PLOT_statistics.R
\name{filter_df_no}
\alias{filter_df_no}
\title{Filter DataFrame by Column Values}
\usage{
filter_df_no(mod_stats, filter_column, filter_by)
}
\arguments{
\item{mod_stats}{A data frame containing model statistics.}

\item{filter_column}{A character vector specifying the columns to filter.}

\item{filter_by}{A list of character vectors containing filter values for each column.}
}
\value{
Returns a filtered data frame if filtering conditions are met; otherwise, NULL.
}
\description{
Filters a data frame based on specified columns and filter values.
}
