% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DD_gene_pairs.R
\name{filter_existing_nodes}
\alias{filter_existing_nodes}
\title{Filter Edges Between Existing Nodes in a Graph}
\usage{
filter_existing_nodes(
  graph,
  edge_df,
  feature1 = "Feature 1",
  feature2 = "Feature 2"
)
}
\arguments{
\item{graph}{An \code{igraph} object containing the network structure.}

\item{edge_df}{A data frame where the first two columns represent
source and target node names, respectively.}
}
\value{
A filtered data frame containing only the edges for which
both nodes exist in the graph.
}
\description{
This function filters an edge data frame to include only those edges
for which both nodes exist in the provided \code{igraph} object.
}
