% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/INPUT_FILTERING_anova_and_trend_test.R
\name{filter_pval}
\alias{filter_pval}
\title{Filter results based on p-value threshold}
\usage{
filter_pval(filtering_res, Pval.th = 0.05)
}
\arguments{
\item{filtering_res}{The filtering results table.}

\item{Pval.th}{The p-value threshold for filtering the results. Default is 0.05.}
}
\value{
A list containing the filtered results and the indices of the filtered rows.
}
\description{
This function filters the  results based on a specified p-value threshold. It returns the filtered ANOVA results
and the indices of the filtered rows.
}
