% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/INPUT_FILTERING_fold_change_limma.R
\name{filter_pval_th}
\alias{filter_pval_th}
\title{Filter Fold Change DataFrame by p-value and fold change threshold}
\usage{
filter_pval_th(fold_change_dataframe, fcPval.th = 0.05, fc.th = 0.58)
}
\arguments{
\item{fold_change_dataframe}{A data frame containing fold change results.}

\item{fcPval.th}{The p-value threshold for filtering (default: 0.05).}

\item{fc.th}{The fold change threshold for filtering (default: 0.58).}
}
\value{
A numeric vector containing the indices of rows that pass the filtering criteria.
}
\description{
This function filters a fold change data frame based on given p-value and fold change thresholds.
}
