% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DD_tpod.R
\name{filter_tpod_acc}
\alias{filter_tpod_acc}
\title{Filter Data Based on tpod accumulation for Given Timepoints}
\usage{
filter_tpod_acc(optimal_models_stats, timepoints_col_variable)
}
\arguments{
\item{optimal_models_stats}{A data frame containing the columns \code{timepoint} and \code{BMD}.
This data frame holds the information about the optimal models for different timepoints.}

\item{timepoints_col_variable}{A character vector specifying the timepoints column variable to analyze.}

\item{plot}{A logical value indicating whether to generate a plot of the accumulation POD
distribution with a vertical line showing the computed tPOD. Default is \code{FALSE}.}
}
\value{
A list containing two data frames:
- \code{before}: data frame with entries where BMD values are less than or equal to the antimode.
- \code{after}: data frame with entries where BMD values are greater than the antimode but less than the 95th percentile.
}
\description{
This function filters data based on the BMD values for specified timepoints. It applies the \code{tpod_accumulation} function
to determine the antimode and then divides the data into two groups:
\itemize{
\item \code{before}: data where BMD values are less than or equal to the antimode.
\item \code{after}: data where BMD values are greater than the antimode but less than the 95th percentile.
}
}
