% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DD_gene_pairs.R
\name{find_best_model_aic}
\alias{find_best_model_aic}
\title{Find the best model based on AIC (Akaike Information Criterion)
This function finds the best model based on the Akaike Information Criterion (AIC) among a list of models.}
\usage{
find_best_model_aic(model_list)
}
\arguments{
\item{model_list}{A list containing multiple models to be compared.}
}
\value{
The model from the list that has the lowest AIC value.
}
\description{
Find the best model based on AIC (Akaike Information Criterion)
This function finds the best model based on the Akaike Information Criterion (AIC) among a list of models.
}
