% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DD_fitting_list.R
\name{fitting_list}
\alias{fitting_list}
\title{Fit dose-response models to data dictionary}
\usage{
fitting_list(
  data_dictionary,
  model_list,
  deviation_type = "relative",
  rl = 1.349,
  confidence_interval = 0.95,
  variance_type = "constant",
  significance_level = 0.05,
  is_parallel = FALSE,
  nCores = 2
)
}
\arguments{
\item{data_dictionary}{The data dictionary containing the data frames to fit the models to.}

\item{model_list}{A list of dose-response models to fit.}

\item{deviation_type}{The type of deviation to use for BMD calculations. Default is "relative".}

\item{rl}{The constant value for relative deviation. Default is 1.349.}

\item{confidence_interval}{The confidence level for the BMD calculation. Default is 0.95.}

\item{variance_type}{The type of variance to assume for the models. Default is "constant".}

\item{significance_level}{The significance level for the BMD calculation. Default is 0.05.}

\item{is_parallel}{Whether to perform the fitting in parallel. Default is FALSE.}

\item{nCores}{The number of cores to use for parallel computation. Default is 2.}
}
\value{
A list containing the fitted models.
}
\description{
This function fits dose-response models to a data dictionary using the specified model list. It returns
a list of fitted models.
}
