% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DD_gene_pairs.R
\name{gene_pairs_analysis}
\alias{gene_pairs_analysis}
\title{Perform gene pairs analysis for selected experiments and times
This function performs gene pairs analysis for selected experiments and times,
using the provided filtered_optimal_models, BMD_tab, length_vectors, nCores,
phenoList, doseColID, timeColID, and other_variables_id_col.}
\usage{
gene_pairs_analysis(
  select_experiment,
  select_time,
  filtered_optimal_models,
  BMD_tab,
  length_vectors,
  nCores,
  phenoList,
  doseColID,
  timeColID,
  other_variables_id_col
)
}
\arguments{
\item{select_experiment}{A character vector specifying the selected
experiments for the analysis.}

\item{select_time}{A numeric vector specifying the selected times for the analysis.}

\item{filtered_optimal_models}{A list of filtered optimal models.}

\item{BMD_tab}{The BMD (Benchmark Dose) table.}

\item{length_vectors}{An integer specifying the length of vectors.}

\item{nCores}{An integer specifying the number of cores to use for parallel processing.}

\item{phenoList}{A list containing the phenotype data.}

\item{doseColID}{The column ID for the dose in the phenotype data.}

\item{timeColID}{The column ID for the time in the phenotype data.}

\item{other_variables_id_col}{The column ID for other variables in the phenotype data.}
}
\value{
A list containing gene pairs statistics and the newdata used for the analysis.
}
\description{
Perform gene pairs analysis for selected experiments and times
This function performs gene pairs analysis for selected experiments and times,
using the provided filtered_optimal_models, BMD_tab, length_vectors, nCores,
phenoList, doseColID, timeColID, and other_variables_id_col.
}
