% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DD_model_statistics.R
\name{get_model_stats}
\alias{get_model_stats}
\title{given a model the function creates a named vector with all the model statistics}
\usage{
get_model_stats(model)
}
\arguments{
\item{model}{an object of class bmdx}
}
\value{
model_stats a named vector with statistics
}
\description{
given a model the function creates a named vector with all the model statistics
}
