% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DD_model_filtering.R
\name{inner.check.model}
\alias{inner.check.model}
\title{Check if a model passes the filtering criteria}
\usage{
inner.check.model(
  mod,
  loofth = 0.1,
  lower_bound_th = 0.1,
  upper_bound_th = 0.1,
  bmd_bmdl_th = 20,
  bmdu_bmd_th = 20,
  bmdu_bmdl_th = 40,
  filter_lower_bound = TRUE,
  filter_upper_bound = TRUE,
  filter_by_lack_of_fit = TRUE,
  ratio_filter = TRUE,
  bmd_na_filter = TRUE,
  bmdl_na_filter = TRUE,
  bmdu_na_filter = TRUE,
  ic50_na_filter = TRUE,
  r2_filter = FALSE,
  r2_th = 0.6,
  filter_by_monotonicity = FALSE,
  filter_by_negative_values = TRUE,
  filter_by_unordered_values = TRUE
)
}
\arguments{
\item{mod}{The model to be checked.}

\item{loofth}{The threshold for lack of fit. Default is 0.1.}

\item{lower_bound_th}{The lower bound threshold. Default is 0.1\% of the lowest dose.}

\item{upper_bound_th}{The upper bound threshold. Default is 0.1\% of the highest dose.}

\item{bmd_bmdl_th}{The threshold for the ratio of BMD to BMDL. Default is 20.}

\item{bmdu_bmd_th}{The threshold for the ratio of BMDU to BMD. Default is 20.}

\item{bmdu_bmdl_th}{The threshold for the ratio of BMDU to BMDL. Default is 40.}

\item{filter_lower_bound}{Logical value indicating whether to filter the model based on
the lower bound threshold. Default is TRUE.}

\item{filter_upper_bound}{Logical value indicating whether to filter the model based on
the upper bound threshold. Default is TRUE.}

\item{filter_by_lack_of_fit}{Logical value indicating whether to filter the model based on
lack of fit. Default is TRUE.}

\item{ratio_filter}{Logical value indicating whether to filter the model based on ratios.
Default is TRUE.}

\item{bmd_na_filter}{Logical value indicating whether to filter the model with missing
BMD values. Default is TRUE.}

\item{bmdl_na_filter}{Logical value indicating whether to filter the model with missing
BMDL values. Default is TRUE.}

\item{bmdu_na_filter}{Logical value indicating whether to filter the model with missing
BMDU values. Default is TRUE.}

\item{ic50_na_filter}{Logical value indicating whether to filter the model with missing
IC50 values. Default is TRUE.}

\item{r2_filter}{Logical value indicating whether to filter the model based on R-squared.
Default is FALSE.}

\item{r2_th}{The threshold for R-squared. Default is 0.6.}

\item{filter_by_monotonicity}{Logical value indicating whether to filter the model based on
monotonicity. Default is FALSE.}
}
\value{
TRUE if the model passes the filtering criteria, FALSE otherwise.
}
\description{
This function checks if a model passes the specified filtering criteria. The criteria
include thresholds, lack of fit, ratios, missing values, R-squared, and monotonicity.
The function returns TRUE if the model passes all the criteria, and FALSE otherwise.
}
