% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DD_model_filtering.R
\name{is_strictly_monotonic}
\alias{is_strictly_monotonic}
\title{Check if a fitted model is strictly monotonic}
\usage{
is_strictly_monotonic(fittedModel)
}
\arguments{
\item{fittedModel}{The fitted model to be checked.}
}
\value{
TRUE if the model is strictly monotonic, FALSE otherwise.
}
\description{
This function checks if a fitted model exhibits strict monotonicity over the range of
doses. It returns TRUE if the model is strictly monotonic and FALSE otherwise.
}
