% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DD_gene_pairs.R
\name{make_d3_network}
\alias{make_d3_network}
\title{Create a D3 Network Visualization}
\usage{
make_d3_network(statistics, g, th = 0, positive = TRUE)
}
\arguments{
\item{statistics}{A data frame containing the statistics on the gene-gene correlation data.}

\item{g}{An igraph graph object representing gene-gene interactions.}

\item{th}{A threshold for filtering correlations.}

\item{positive}{Logical, whether to consider positive correlations.}
}
\value{
A list containing the D3 network visualization, a data frame of node statistics, and an igraph graph object.
}
\description{
This function generates a D3 network visualization based on statistical data and a graph structure.
}
