% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PLOT_statistics.R
\name{make_empty_plot}
\alias{make_empty_plot}
\title{Create Empty ggplot2 Plot
Creates an empty ggplot2 plot with a void theme and no x-label.}
\usage{
make_empty_plot()
}
\value{
Returns an empty ggplot2 plot.
}
\description{
Create Empty ggplot2 Plot
Creates an empty ggplot2 plot with a void theme and no x-label.
}
