% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DD_model_filtering.R
\name{model_filtering}
\alias{model_filtering}
\title{Filter a list of fitted models based on various criteria}
\usage{
model_filtering(
  fitted_models,
  loofth = 0.1,
  lower_bound_th = 0.1,
  upper_bound_th = 0.1,
  bmd_bmdl_th = 20,
  bmdu_bmd_th = 20,
  bmdu_bmdl_th = 40,
  filter_lower_bound = TRUE,
  filter_upper_bound = TRUE,
  filter_by_lack_of_fit = TRUE,
  ratio_filter = TRUE,
  bmd_na_filter = TRUE,
  bmdl_na_filter = TRUE,
  bmdu_na_filter = TRUE,
  ic50_na_filter = TRUE,
  r2_filter = FALSE,
  r2_th = 0.6,
  filter_by_monotonicity = FALSE,
  filter_by_negative_values = TRUE,
  filter_by_unordered_values = TRUE
)
}
\arguments{
\item{fitted_models}{A list of fitted models to be filtered.}

\item{loofth}{The threshold for lack of fit. Default is 0.1.}

\item{lower_bound_th}{The lower bound threshold. Default is 0.1\% of the lowest dose.}

\item{upper_bound_th}{The upper bound threshold. Default is 0.1\% of the highest dose.}

\item{bmd_bmdl_th}{The threshold for the ratio of BMD to BMDL. Default is 20.}

\item{bmdu_bmd_th}{The threshold for the ratio of BMDU to BMD. Default is 20.}

\item{bmdu_bmdl_th}{The threshold for the ratio of BMDU to BMDL. Default is 40.}

\item{filter_lower_bound}{Logical value indicating whether to filter models based on
the lower bound threshold. Default is TRUE.}

\item{filter_upper_bound}{Logical value indicating whether to filter models based on
the upper bound threshold. Default is TRUE.}

\item{filter_by_lack_of_fit}{Logical value indicating whether to filter models based on
lack of fit. Default is TRUE.}

\item{ratio_filter}{Logical value indicating whether to filter models based on ratios.
Default is TRUE.}

\item{bmd_na_filter}{Logical value indicating whether to filter models with missing
BMD values. Default is TRUE.}

\item{bmdl_na_filter}{Logical value indicating whether to filter models with missing
BMDL values. Default is TRUE.}

\item{bmdu_na_filter}{Logical value indicating whether to filter models with missing
BMDU values. Default is TRUE.}

\item{ic50_na_filter}{Logical value indicating whether to filter models with missing
IC50 values. Default is TRUE.}

\item{r2_filter}{Logical value indicating whether to filter models based on R-squared.
Default is FALSE.}

\item{r2_th}{The threshold for R-squared. Default is 0.6.}

\item{filter_by_monotonicity}{Logical value indicating whether to filter models based on
monotonicity. Default is FALSE.}

\item{filter_by_negative_values}{Logical value indicating wheather to filter models based on negative estimated effective doses}

\item{filter_by_unordered_values}{Logival value indicating wheather to filter models for which BMDL<BMD<BMDU is not true}
}
\value{
A filtered list of models that pass the specified criteria.
}
\description{
This function filters a list of fitted models based on specified criteria, such as
thresholds, lack of fit, ratios, missing values, R-squared, and monotonicity.
The function returns a filtered list of models that pass the specified criteria.
}
