% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/INPUT_FILTERING_anova_and_trend_test.R
\name{perform_anova}
\alias{perform_anova}
\title{Perform ANOVA on data dictionary}
\usage{
perform_anova(
  data_dictionary,
  anovaAdjustment = "Nominal",
  anovaPval.th = 0.05,
  anovaCores = 1,
  x = "dose",
  y = "expr",
  other_variables_id_col = NULL
)
}
\arguments{
\item{data_dictionary}{The data dictionary containing the data frames to perform ANOVA on.}

\item{anovaAdjustment}{The adjustment method for p-values. Options include "Nominal" (no adjustment) and
methods supported by the \code{p.adjust} function. Default is "Nominal".}

\item{anovaPval.th}{The significance threshold for filtering the ANOVA results. Default is 0.05.}

\item{anovaCores}{The number of cores to use for parallel computation. Default is 1.}

\item{x}{The column name for the independent variable. Default is "dose".}

\item{y}{The column name for the dependent variable. Default is "expr".}

\item{other_variables_id_col}{Additional column names for other variables to include in the ANOVA.}
}
\value{
anova_res_list A list containing the ANOVA results dataframe, filtered data dictionary, unfiltered ANOVA results dataframe, and a plot list.
}
\description{
This function performs ANOVA on a data dictionary, computes p-values, performs adjustment if specified,
filters the results based on a significance threshold, and returns the ANOVA results and filtered data
dictionary.
}
