% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/INPUT_FILTERING_fold_change_limma.R
\name{perform_differential_expression_analysis_filtering}
\alias{perform_differential_expression_analysis_filtering}
\title{Perform Differential Expression Analysis}
\usage{
perform_differential_expression_analysis_filtering(
  data_dictionary,
  experimental_data,
  phTable,
  time_point_variable,
  dose_variable,
  samples_variable,
  fcAdjustment = "Nominal",
  fcPval.th = 0.05,
  fc.th = 1.5,
  nCores = 1,
  x = "dose",
  y = "expr",
  other_variables_id_col = NULL
)
}
\arguments{
\item{data_dictionary}{A list of data frames representing the data dictionary.}

\item{experimental_data}{A list of data frames containing the experimental data.}

\item{phTable}{A list of data frames representing the phenotype table.}

\item{time_point_variable}{The name of the variable representing time points.}

\item{dose_variable}{The name of the variable representing doses.}

\item{samples_variable}{The name of the variable representing samples.}

\item{fcAdjustment}{The type of fold change adjustment (default: "Nominal").
Options are "none" or "whatever_other_adjustment".}

\item{fcPval.th}{The p-value threshold for fold change filtering (default: 0.05).}

\item{fc.th}{The fold change threshold (default: 1.5).}

\item{nCores}{The number of cores for parallel processing (default: 1).}

\item{x}{The name of the x-axis variable for plotting (default: "dose").}

\item{y}{The name of the y-axis variable for plotting (default: "expr").}

\item{other_variables_id_col}{A vector of other variable names used for filtering.}
}
\value{
A list containing the fold change results (filtered and unfiltered),
the filtered data dictionary, and the filtered fold change results.
}
\description{
This function filter the data based on their differential analysis
}
