% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/INPUT_FILTERING_anova_and_trend_test.R
\name{perform_trend_test}
\alias{perform_trend_test}
\title{Perform the trend test on multiple datasets.}
\usage{
perform_trend_test(
  data_dictionary,
  trendAdjustment = "Nominal",
  trendPval.th = 0.05,
  trendCores = 1,
  x = "dose",
  y = "expr",
  other_variables_id_col = NULL
)
}
\arguments{
\item{data_dictionary}{A list of datasets, each stored as a data frame in the list.}

\item{trendAdjustment}{The method for adjusting p-values. Default is "Nominal".
Available options include "Bonferroni", "Holm", "Hochberg",
"BH", "BY", "fdr", "none".}

\item{trendPval.th}{The threshold for filtering p-values. Default is 0.05.}

\item{trendCores}{The number of CPU cores to use for parallel processing. Default is 1.}

\item{x}{The column name representing the predictor variable (independent variable) in each dataset.}

\item{y}{The column name representing the response variable (dependent variable) in each dataset.}

\item{other_variables_id_col}{The column name containing the identifier for other variables.
Default is NULL.}
}
\value{
A list containing the trend test results and filtered datasets.
}
\description{
This function performs the trend test on a list of datasets stored in the
\code{data_dictionary}. The trend test is performed using the \code{mk.test}
function from the \code{trend} package. It calculates the trend p-value for
each dataset and adjusts the p-values using the specified \code{trendAdjustment}
method. It also performs filtering based on the \code{trendPval.th} threshold.
}
