% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DD_plot_bmd_bmdl_bmdu_set_of_genes.R
\name{plot_bmd_bmdl_bmdu_set_of_genes}
\alias{plot_bmd_bmdl_bmdu_set_of_genes}
\title{Plot BMD, BMDL, and BMDU for a Set of Genes}
\usage{
plot_bmd_bmdl_bmdu_set_of_genes(
  BMDFilMat,
  gi,
  enrich_ppi_info = TRUE,
  gene_id_type = "ENSEMBL",
  organism = "human"
)
}
\arguments{
\item{BMDFilMat}{A data frame containing BMD values for multiple genes.}

\item{gi}{A vector of gene names for which to plot BMD values.}
}
\value{
A plot showing BMD, BMDL, and BMDU values for the selected set of genes.
}
\description{
This function generates a plot showing BMD, BMDL, and BMDU values for a given set of genes.
}
