% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MODEL_plot_model.R
\name{plot_bmdx}
\alias{plot_bmdx}
\title{Plot BMD (Benchmark Dose) Model}
\usage{
plot_bmdx(
  model,
  cex = 6,
  confidence_interval = 0.95,
  plot_ic50 = FALSE,
  xlim_u = NULL,
  title_label = "title"
)
}
\arguments{
\item{model}{The fitted BMD model.}

\item{cex}{The size of points in the plot (default: 6).}

\item{confidence_interval}{The confidence interval level (default: 0.95).}

\item{title_label}{The title of the plot (default: "title").}

\item{x_pos}{The position of the x-axis (default: 15).}

\item{y_pos_th}{The position of the y-axis threshold (default: 0.85).}
}
\value{
A plot displaying the BMD model with confidence intervals and other related data points.
}
\description{
This function generates a plot for a Benchmark Dose (BMD) model.
}
