% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/INPUT_FILTERING_anova_and_trend_test.R
\name{plot_filtering_pie_chart}
\alias{plot_filtering_pie_chart}
\title{Plot a pie chart of filtering results}
\usage{
plot_filtering_pie_chart(tab_unfiltered, Pval.th, title = "")
}
\arguments{
\item{tab_unfiltered}{The unfiltered table containing the filtering p-values.}

\item{Pval.th}{The p-value threshold for determining variable significance.}

\item{title}{Title of the plot. It is a string. default = ""}
}
\value{
A pie chart visualization of the filtering results.
}
\description{
This function generates a pie chart to visualize the filtering results. The pie chart shows the distribution of variables
and non-variables based on the filtering p-values in the table. The function takes the table and an
p-value threshold as input.
}
