% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DD_gene_pairs.R
\name{plot_gene_pairs}
\alias{plot_gene_pairs}
\title{Plot gene pairs comparison}
\usage{
plot_gene_pairs(mod_i, mod_j, newdata, main = "", feat1 = "g1", feat2 = "g2")
}
\arguments{
\item{mod_i}{The model for gene pair 1.}

\item{mod_j}{The model for gene pair 2.}

\item{newdata}{A data frame containing dose information for new observations.}

\item{main}{A character string specifying the main title of the plot
(default is an empty string).}

\item{feat1}{A character string specifying the label for gene pair 1 in the
plot (default is "g1").}

\item{feat2}{A character string specifying the label for gene pair 2 in the plot
(default is "g2").}
}
\value{
A ggplot object showing the comparison of gene pairs.
}
\description{
This function plots the comparison of two gene pairs given the models mod_i
and mod_j and newdata containing dose information for new observations.
}
