% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PLOT_statistics.R
\name{plot_histogram}
\alias{plot_histogram}
\title{Plot Histogram and Density}
\usage{
plot_histogram(
  mod_stats,
  y_val = "BMD",
  color_by = NULL,
  group_by = NULL,
  group_by2 = NULL,
  filter_column = NULL,
  filter_by = NULL,
  alpha_th = 0.5
)
}
\arguments{
\item{mod_stats}{A data frame containing the data for plotting.}

\item{y_val}{The variable to be plotted on the y-axis (numeric).}

\item{color_by}{The categorical variable used for color-coding bars.}

\item{group_by}{The grouping variable for additional data segmentation.}

\item{group_by2}{The second grouping variable for further segmentation.}

\item{filter_column}{The column name for filtering the data.}

\item{filter_by}{The values used for filtering the data.}

\item{alpha_th}{The transparency level of the density plot (ranges between 0-1).}
}
\value{
A ggplot2 histogram and density plot visualization.
}
\description{
Generate a histogram and density plot visualization based on provided data and variables.
}
