% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PLOT_statistics.R
\name{plot_pie_chart}
\alias{plot_pie_chart}
\title{this function allows to plot the histogram of one numeric column of the model statistics}
\usage{
plot_pie_chart(
  mod_stats,
  category = "Model",
  group_by = NULL,
  group_by2 = NULL,
  filter_column = NULL,
  filter_by = NULL
)
}
\arguments{
\item{mod_stats}{A data frame containing the data for plotting.}

\item{category}{The category variable to be used for the pie chart sectors.}

\item{group_by}{The grouping variable for additional data segmentation.}

\item{group_by2}{The second grouping variable for further segmentation.}

\item{filter_column}{The column name for filtering the data.}

\item{filter_by}{The values used for filtering the data.}
}
\value{
a ggplot object
}
\description{
this function allows to plot the histogram of one numeric column of the model statistics
}
