% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PLOT_statistics.R
\name{plot_scatter}
\alias{plot_scatter}
\title{Plot Scatter Plot}
\usage{
plot_scatter(
  mod_stats,
  x_val = "BMDL",
  y_val = "BMD",
  color_by = "Model",
  group_by = NULL,
  group_by2 = NULL,
  filter_column = NULL,
  filter_by = NULL
)
}
\arguments{
\item{mod_stats}{A data frame containing the data for plotting.}

\item{x_val}{The variable to be plotted on the x-axis.}

\item{y_val}{The variable to be plotted on the y-axis.}

\item{color_by}{The categorical variable used for color-coding points.}

\item{group_by}{The grouping variable for additional data segmentation.}

\item{group_by2}{The second grouping variable for further segmentation.}

\item{filter_column}{The column name for filtering the data.}

\item{filter_by}{The values used for filtering the data.}
}
\value{
A ggplot2 scatter plot visualization.
}
\description{
Generate a scatter plot visualization based on provided data and variables.
}
