% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DD_point_of_departure.R
\name{point_of_departure}
\alias{point_of_departure}
\title{Given a model, this function estimates the Benchmark Dose (BMD),
Benchmark Dose Lower Confidence Limit (BMDL), Benchmark Dose Upper Confidence
Limit (BMDU), and AC50 values.}
\usage{
point_of_departure(
  model,
  deviation_type = "standard",
  rl = 1.349,
  confidence_interval = 0.95
)
}
\arguments{
\item{model}{The model object representing the dose-response relationship.}

\item{deviation_type}{Character string specifying the type of deviation from
the fitted model to use for BMD calculation. Default is "standard".}

\item{rl}{The relative level used to calculate the BMD. Default is 1.349.}

\item{confidence_interval}{The confidence level for the confidence interval.
Default is 0.95.}
}
\value{
A modified model object with additional attributes for BMDL, BMDU, and AC50.
If an error occurs during the estimation, NA values are assigned to BMDL and BMDU.
}
\description{
Given a model, this function estimates the Benchmark Dose (BMD),
Benchmark Dose Lower Confidence Limit (BMDL), Benchmark Dose Upper Confidence
Limit (BMDU), and AC50 values.
}
