% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/INPUT_FILTERING_anova_and_trend_test.R
\name{pval_adjust}
\alias{pval_adjust}
\title{Adjust p-values in filtering results}
\usage{
pval_adjust(filtering_res, adjustment_method = "fdr")
}
\arguments{
\item{filtering_res}{The filtering results table.}

\item{adjustment_method}{The adjustment method for p-values. Options include "Nominal" (no adjustment) and
methods supported by the \code{p.adjust} function. Default is "fdr" (false discovery rate).}
}
\value{
The filtering results table with adjusted p-values and the column "usedFilteringPval" indicating the
p-values used for filtering.
}
\description{
This function adjusts the p-values in the filtering results using a specified adjustment method.
}
