% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/INPUT_read_excel_allsheets.R
\name{read_excel_allsheets}
\alias{read_excel_allsheets}
\title{Read all sheets from an Excel file.}
\usage{
read_excel_allsheets(
  filename,
  tibble = FALSE,
  first_col_as_rownames = FALSE,
  is_rnaseq_raw_count = FALSE,
  check_numeric = TRUE,
  na = "NA"
)
}
\arguments{
\item{filename}{The path to the Excel file.}

\item{tibble}{If \code{TRUE}, the output data frames will be converted to tibbles.
Default is \code{FALSE}.}

\item{first_col_as_rownames}{If \code{TRUE}, the first column of each sheet will be used as row names.
Default is \code{FALSE}.}

\item{is_rnaseq_raw_count}{If \code{TRUE}, the data is assumed to be RNA-Seq raw counts, and it will
be converted to log2 counts using the \code{limma::voom} function.
Default is \code{FALSE}.}
}
\value{
A list of data frames, one for each sheet in the Excel file.
}
\description{
This function reads all sheets from an Excel file specified by the \code{filename}.
It returns the data as a list of data frames, one for each sheet.
}
