% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PLOT_compare_experiments.R
\name{scale_numbers}
\alias{scale_numbers}
\title{Scale Numbers to the Range 0-1
This function scales numeric values to the range 0-1 by dividing each value by the maximum value in the input vector.}
\usage{
scale_numbers(x)
}
\arguments{
\item{x}{A numeric vector to be scaled.}
}
\value{
A numeric vector with scaled values in the range 0-1.
}
\description{
Scale Numbers to the Range 0-1
This function scales numeric values to the range 0-1 by dividing each value by the maximum value in the input vector.
}
