% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DD_select_optimal_models.R
\name{select_optimal_models}
\alias{select_optimal_models}
\title{Select optimal models from a list of current models}
\usage{
select_optimal_models(
  current_models,
  method = "AIC",
  time_col_id,
  optional_col_ids = NULL,
  nCores = 1
)
}
\arguments{
\item{current_models}{The list of current models.}

\item{method}{The method for selecting the optimal models. Supported values are "AIC" and "Model average".}

\item{time_col_id}{The identifier for the time column in the model statistics table.}

\item{optional_col_ids}{Optional identifiers for additional columns in the model statistics table.}

\item{nCores}{The number of CPU cores to use for parallel computation.}
}
\value{
A list containing the optimal models and the computed model statistics table.
}
\description{
This function selects the optimal models from a list of current models based on the specified method. The supported
methods are "AIC" (Akaike Information Criterion) and "Model average". For the "AIC" method, the model with the lowest
AIC value is selected for each dataset. For the "Model average" method, if average models are already being computed,
the average model is selected; otherwise, the first model in the list is selected. The function returns a list
containing the optimal models and a table of computed model statistics.
}
