% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sym_human_gene_graph}
\alias{sym_human_gene_graph}
\title{sym_human_gene_graph}
\format{
An object of class \code{igraph} of length 18932.
}
\source{
\url{https://doi.org/10.1093/bib/bbaa417}
}
\usage{
sym_human_gene_graph
}
\description{
Graph whose nodes are Human genes (Gene Symbols). Edges represent protein-protein interactions (PPIs).
}
\keyword{datasets}
