% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sym_mouse_gene_graph_tf}
\alias{sym_mouse_gene_graph_tf}
\title{sym_mouse_gene_graph_tf}
\format{
An object of class \code{igraph} of length 5062.
}
\source{
\url{https://doi.org/10.1093/bib/bbaa417}
}
\usage{
sym_mouse_gene_graph_tf
}
\description{
Graph whose nodes are Mouse genes (Gene Symbols). Edges represent regulatory interactions between transcription factor (TF) and genes.
}
\keyword{datasets}
