% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sym_rat_gene_graph_mirna}
\alias{sym_rat_gene_graph_mirna}
\title{sym_rat_gene_graph_mirna}
\format{
An object of class \code{igraph} of length 24.
}
\source{
\url{https://doi.org/10.1093/bib/bbaa417}
}
\usage{
sym_rat_gene_graph_mirna
}
\description{
Graph whose nodes are Rat genes (Gene Symbols). Edges represent regulatory interactions between miRNA and genes.
}
\keyword{datasets}
