% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DD_tpod.R
\name{tpod_accumulation}
\alias{tpod_accumulation}
\title{Calculate tPOD Using an Accumulation-Based Approach}
\usage{
tpod_accumulation(pod_vector, plot = FALSE, bw_adjust = 3)
}
\arguments{
\item{pod_vector}{A numeric vector representing the POD (Point of Departure) values.}

\item{plot}{Logical. If \code{TRUE}, a plot is generated showing the sorted data, the fitted curve,
a vertical line at the antimode, and another vertical line at the estimated tPOD.}

\item{...}{Additional arguments passed to \code{get_mode_antimode}.}
}
\value{
A single numeric value representing the tPOD, identified by the Kneedle method on the fitted cumulative distribution.
}
\description{
This function calculates the tPOD (Point of Departure) by:
\enumerate{
\item Identifying the mode and antimode of the input distribution (via \code{get_mode_antimode}).
\item Filtering the data to values below the antimode.
\item Fitting a shape-constrained additive model (via \code{scam}).
\item Applying the Kneedle algorithm to the smoothed cumulative sum to detect the knee point.
}
}
\details{
\enumerate{
\item \strong{Sorting}: The function sorts \code{pod_vector} in increasing order.
\item \strong{Accumulation}: It computes the cumulative sum of the sorted values (y) and the log-transformed x-axis (log10 of the sorted values).
\item \strong{Filtering}: Using the estimated \code{antimode} from \code{get_mode_antimode}, data points above the antimode are removed.
\item \strong{Fitting}: A shape-constrained additive model (\code{scam}) is then fitted to the filtered data to smooth the cumulative sums.
\item \strong{Knee Detection}: The Kneedle algorithm (\code{kneedle}) is applied to the fitted curve to detect the knee, which is reported as the tPOD.
\item \strong{Optional Plot}: If \code{plot = TRUE}, the sorted data, fitted curve, and vertical lines marking the antimode and tPOD will be displayed.
}
}
