% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DD_tpod.R
\name{tpod_computation}
\alias{tpod_computation}
\title{Compute tPOD using various methods and plot the result}
\usage{
tpod_computation(
  model_stats,
  tpod_method,
  percentile,
  pod_value = "BMD",
  lowest_method = "lowest",
  bw_adjust = 3,
  ratio_threshold = 1.66,
  plot = FALSE,
  robust = F,
  rank = 1
)
}
\arguments{
\item{model_stats}{A data frame or list containing the model statistics, including the
POD values (e.g., BMD, BMDL, or BMDU) from which tPOD will be computed.}

\item{tpod_method}{A string indicating the method to compute tPOD. Options include:
"percentile", "mean", "first_mode", "lowest", or "accumulation".}

\item{percentile}{A numeric value between 0 and 1 representing the desired percentile
(used if \code{tpod_method} is "percentile").}

\item{pod_value}{A string indicating which POD value to use from the model statistics.
Possible values are "BMD", "BMDL", and "BMDU" (default is "BMD").}
}
\value{
A numeric value representing the computed tPOD, according to the selected method.
}
\description{
This function computes the tPOD (Point of Departure) using different methods
such as percentile, mean, first mode, lowest, or accumulation After calculating the tPOD,
it generates a plot displaying the resulting tPOD value on the accumulation POD distribution.
}
\details{
The function computes the tPOD using one of the following methods:
\itemize{
\item \strong{percentile}: Computes tPOD at the specified percentile of the POD values.
\item \strong{mean}: Computes the tPOD as the mean value of the POD values.
\item \strong{first_mode}: Computes the tPOD using the first mode (most frequent value) of the POD distribution.
\item \strong{lowest}: Computes the tPOD as the lowest value of the POD distribution.
\item \strong{accumulation}: Computes the tPOD using the accumulation method.
}

A plot is also generated showing the accumulation plot of the POD distribution with a vertical line indicating
the tPOD value.
}
