% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DD_tpod.R
\name{tpod_first_mode}
\alias{tpod_first_mode}
\title{Calculate the first mode value using kernel density estimation}
\usage{
tpod_first_mode(pod_vector)
}
\arguments{
\item{pod_vector}{A numeric vector for which the mode is to be calculated.}
}
\value{
A numeric value representing the mode of the input vector.
}
\description{
This function computes the mode of a given vector using kernel density estimation (KDE).
The mode is the value that has the highest density in the distribution of the data.
}
\details{
The function uses the kernel density estimation technique to estimate the probability density function of the given vector.
It then identifies the value that corresponds to the highest density, which is returned as the mode.
This approach is useful when the distribution is not unimodal or not normally distributed.
}
