% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DD_tpod.R
\name{tpod_lowest}
\alias{tpod_lowest}
\title{Calculate the lowest value of a numeric vector, excluding NA values}
\usage{
tpod_lowest(pod_vector, lowest_method, ratio_threshold, rank)
}
\arguments{
\item{pod_vector}{A numeric vector for which the lowest value is to be calculated.}
}
\value{
A numeric value representing the lowest value of the input vector, excluding \code{NA}s.
}
\description{
This function computes the lowest value in the given numeric vector, excluding any \code{NA} values.
}
\details{
The function uses the \code{min} function with the \code{na.rm = TRUE} argument to exclude \code{NA} values
when computing the minimum. It returns the lowest value as a numeric result.
}
