% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DD_tpod.R
\name{tpod_mean}
\alias{tpod_mean}
\title{Calculate the Mean of a POD Vector}
\usage{
tpod_mean(pod_vector)
}
\arguments{
\item{pod_vector}{A numeric vector containing POD values.}
}
\value{
A numeric value representing the mean of the POD vector.
}
\description{
This function computes the mean of a given POD (Point of Departure) vector, excluding NA values.
}
