% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DD_tpod.R
\name{tpod_percentile}
\alias{tpod_percentile}
\title{Calculate the Percentile of a POD Vector}
\usage{
tpod_percentile(pod_vector, percentile)
}
\arguments{
\item{pod_vector}{A numeric vector containing POD values.}

\item{percentile}{A numeric value representing the desired percentile (should be within \link{0,1}).}
}
\value{
A numeric value representing the calculated percentile of the POD vector.
}
\description{
This function computes a specified percentile of a given POD (Point of Departure) vector.
}
