% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DD_tpod.R
\name{tpod_plot}
\alias{tpod_plot}
\title{Generate an accumulation plot with multiple tPOD methods}
\usage{
tpod_plot(
  pod_vector,
  tpod_method,
  pod_value,
  tPOD,
  xlog = FALSE,
  subtitle = ""
)
}
\arguments{
\item{pod_vector}{A numeric vector representing the POD (Point of Departure) values.}

\item{tpod_method}{A character vector containing the names of the tPOD methods used
to calculate the points of Departure.}

\item{pod_value}{A string that represents the value used to calculate the tPODs,
for example, "BMD" or "BMDL" or "BMDU".}

\item{tPOD}{A numeric vector containing the points of Departure (tPODs)
corresponding to each method. The length of this vector must match the length of
the \code{tpod_method} vector.}

\item{xlog}{A logical value indicating whether to apply a logarithmic transformation
to the x-axis (default is \code{FALSE}).}
}
\value{
A ggplot object representing the cumulative sum plot with vertical lines for tPODs.
}
\description{
This function generates a cumulative sum plot of the given POD vector, and includes
vertical lines indicating the tPODs (points of Departure) for multiple methods.
The plot can display multiple tPOD values if they all have the same length.
The x-axis can optionally be transformed to a logarithmic scale.
}
\details{
The function generates an accumulation plot of the sorted POD vector and adds vertical
lines at the locations of the provided tPOD values. The plot can display multiple tPOD
values (with different methods) if the length of \code{tPOD} and \code{tpod_method} are equal.
The x-axis can optionally be displayed in logarithmic scale.
}
