% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PLOT_compare_experiments.R
\name{upset_plot}
\alias{upset_plot}
\title{Create an UpSet Plot}
\usage{
upset_plot(
  mod_stats,
  rel_variable = "Experiment",
  group_by = NULL,
  other_variables = NULL,
  filter_column = c("Model"),
  filter_by = list(c("linear")),
  nintersects = 3,
  group.by = "degree",
  order.by = "degree",
  text.scale = c(1.5, 1.5, 1.2, 1.2, 1.5, 1.5)
)
}
\arguments{
\item{mod_stats}{A data frame containing the model statistics and gene information.}

\item{rel_variable}{The name of the variable representing the experiments or conditions.}

\item{group_by}{The name of the variable to group the data for generating UpSet plots.}

\item{other_variables}{Additional variables used for plotting, if any.}

\item{filter_column}{The name of the column to filter the data.}

\item{filter_by}{A list of filter values to apply on the specified filter_column.}

\item{nintersects}{The number of intersections to show in the UpSet plot.}

\item{group.by}{The variable used to group the intersections (e.g., "degree" or "freq").}

\item{order.by}{The variable used to order the intersections (e.g., "frequency", or "degree").}
}
\value{
An UpSet plot displaying the intersections of sets of genes across different experiments or conditions.
}
\description{
This function generates an UpSet plot for the given data, which displays the intersections of sets of genes across different experiments or conditions.
}
